+++ src/main/java/com/academy/mortgage/controllers/ApplicationsController.java
@@ -1,5 +1,6 @@
-package com.academy.mortgage.applications;
+package com.academy.mortgage.controllers;
 
+import com.academy.mortgage.repositories.ApplicationsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
