+++ src/main/java/com/academy/mortgage/controllers/UsersController.java
@@ -1,10 +1,11 @@
-package com.academy.mortgage.users;
+package com.academy.mortgage.controllers;
 
+import com.academy.mortgage.repositories.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-class UsersController {
+public class UsersController {
     @Autowired
     UsersRepository usersRepository;
 }
