--- src/main/java/com/academy/mortgage/security/SecurityConfig.java
@@ -1,8 +1,0 @@
-package com.academy.mortgage.security;
-
-public class SecurityConfig {
-
-}
-
-
-
