+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.datasource.url=jdbc:postgresql://dpg-cgqrkviut4mde4or9u00-a.frankfurt-pos
 #spring.datasource.url=jdbc:${DATABASE_ADDRESS}
 spring.jpa.show-sql=true
 spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.sql.init.mode=always
